+++ Source/Lib/Color/Aul2ColorUtils.pas
@@ -1,4 +1,4 @@
-unit Aul2ColorUtils;
+﻿unit Aul2ColorUtils;
 
 interface